@@ -1,6 +1,6 @@
 module github.com/VitalyKondratiev/gtrack
 
-go 1.17
+go 1.21
 
 require (
 	github.com/Jeffail/gabs v1.4.0
